move related functions to own modules
